git-svn-id: https://svn.kuali.org/repos/student/ks-core/branches/ks-core-dev@3144 0ac02d80-fb3f-11dc-a574-c57244a850eb

+++ core-impl/src/main/java/org/kuali/student/core/dao/impl/AbstractCrudDaoImpl.java
@@ -1,17 +1,29 @@
 package org.kuali.student.core.dao.impl;
 
 import java.util.List;
+import java.util.Locale;
 
 import javax.persistence.EntityManager;
 import javax.persistence.Query;
 
 import org.kuali.student.core.dao.CrudDao;
 import org.kuali.student.core.entity.AttributeDef;
+import org.kuali.student.core.exceptions.DoesNotExistException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
 
 public abstract class AbstractCrudDaoImpl implements CrudDao {
 
 	protected EntityManager em;
-	
+	@Autowired
+	private MessageSource messageSource;
+	private Locale locale;
+
+	public void setMessageSource (MessageSource messageSource) {
+		this.messageSource = messageSource;
+		locale = Locale.getDefault();
+	}
+
 	public EntityManager getEm() {
 		return em;
 	}
@@ -22,16 +34,21 @@ public abstract class AbstractCrudDaoImpl implements CrudDao {
 
 
 	@Override
-	public <T> T fetch(Class<T> clazz, String key) {
-		return em.find(clazz, key);
+	public <T> T fetch(Class<T> clazz, String key) throws DoesNotExistException {
+		T entity = em.find(clazz, key);
+		if (entity == null) {
+			String msg = messageSource.getMessage("AbstractCrudDaoImpl.noEntityForKey", new Object[] {key, clazz}, locale);
+			throw new DoesNotExistException(msg);
+		}
+		return entity;
 	}
-	
+
 	@SuppressWarnings("unchecked")
 	@Override
 	public <T> List<T> find(Class<T> clazz){
-		
+
 		String className = clazz.getSimpleName();
-		
+
 		Query q = em.createQuery("SELECT x FROM "+className+" x");
 		return (List<T>) q.getResultList();
 	}
@@ -41,7 +58,7 @@ public abstract class AbstractCrudDaoImpl implements CrudDao {
 	public <T extends AttributeDef> T fetchAttributeDefByName(Class<T> clazz, String attributeName) {
 
 		String className = clazz.getSimpleName();
-		
+
 		Query q = em.createQuery("SELECT attrDef FROM "+className+" attrDef WHERE attrDef.name = :attributeName");
 		q.setParameter("attributeName", attributeName);
 
@@ -55,16 +72,20 @@ public abstract class AbstractCrudDaoImpl implements CrudDao {
 	}
 
 	@Override
-	public <T> void delete(Class<T> clazz, String key) {
+	public <T> void delete(Class<T> clazz, String key) throws DoesNotExistException {
 		T entity = em.find(clazz, key);
+		if (entity == null) {
+			String msg = messageSource.getMessage("AbstractCrudDaoImpl.noEntityForKey", new Object[] {key, clazz}, locale);
+			throw new DoesNotExistException(msg);
+		}
 		em.remove(entity);
 	}
 
 	@Override
 	public void delete(Object entity) {
-		em.remove(entity);	
+		em.remove(entity);
 	}
-	
+
 	@Override
 	public <T> T update(T entity) {
 		T updated =  em.merge(entity);
